+++ app/Http/Controllers/Api/SupplierController.php
@@ -117,7 +117,9 @@ class SupplierController extends Controller
             $data['photo'] = $image_url;
             $img = DB::table('suppliers')->where('id',$id)->first();
             $image_path = $img->photo;
-            $done = unlink($image_path);
+            if($img->photo){
+                $done = unlink($image_path);
+            }
             $user  = DB::table('suppliers')->where('id',$id)->update($data);
          }
